Minor change
Former-commit-id: d18c06b45f043a252aa719e2dda2d1896dd8452b

+++ include/preProcess.hpp
@@ -102,7 +102,7 @@ void computeKmerFrequency(std::vector<T>& localvector, MPI_Comm comm = MPI_COMM_
  * @tparam[in] filterbyMedian     Filter if median exceeds "medianCutOff" in the Pc bucket
  * @tparam[in] filterbyMax        Filter if maximum kmer frequence exceeds "frequencyCutOff" in the Pc bucket (which is a read during preprocessing)
  */
-template <unsigned int tmpLayer, unsigned int readIdLayer, bool filterbyMedian, bool filterbyMax, uint8_t medianCutOff = HIST_EQ_THRESHOLD, uint8_t frequencyCutOff = KMER_FREQ_THRESHOLD, typename T>
+template <unsigned int tmpLayer, unsigned int readIdLayer, bool filterbyMedian, bool filterbyMax, uint32_t medianCutOff = HIST_EQ_THRESHOLD, uint32_t frequencyCutOff = KMER_FREQ_THRESHOLD, typename T>
 void updateReadFilterFlags(std::vector<T>& localvector, std::vector<bool>& readFilterFlags, uint32_t firstReadId, MPI_Comm comm = MPI_COMM_WORLD)
 {
   //Know my rank
